% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon_stop_codon_remover.R
\name{remove_stop_codons}
\alias{remove_stop_codons}
\title{remove_stop_codons}
\usage{
remove_stop_codons(fasta, fileoutprefix)
}
\arguments{
\item{fasta}{Input file in fasta format}

\item{fileoutprefix}{The base name to be used in the output files}
}
\value{
None, writes FASTA files with and without stop codons.
}
\description{
remove_stop_codons
}
